client/requirements.txt
grpcio>=1.71.2,<2.0
protobuf>=4.25.3,<6.0
rich==13.9.4
